Adding diffuse and ambient lighting, some cleanup

+++ src/Camera.cpp
@@ -11,7 +11,7 @@ using namespace Eigen;
 
 // Coordinate system: OpenGL
 
-Camera::Camera( const Vector3d &origin, const float &fov, const Vector2d &resolution ):
+Camera::Camera( const Vector3d &origin, const float &fov, const Vector2i &resolution ):
         _origin(origin), _fov(fov), _resolution(resolution) {};
 
 // Given a viewport size (image resolution), origin, and focal point, compute
@@ -23,12 +23,12 @@ std::map< std::vector<int>, Ray > Camera::getAllRays()
   
   // For each pixel, get its location in world coordinates and compute ray
   // from origin to pixelpoint.
-  double aspectRatio = _resolution.x()/_resolution.y();
+  double aspectRatio = _resolution.x()/static_cast<double>(_resolution.y());
   double scale = tan(_fov * (M_PI/180.) * 0.5);
 
-  for( int x = 0; x < int(_resolution.x()); x++ )
+  for( int x = 0; x < _resolution.x(); x++ )
   {
-    for( int y = 0; y < int(_resolution.y()); y++ )
+    for( int y = 0; y < _resolution.y(); y++ )
     { 
       // Raster --> Screen --> Camera Coordinate
       double pixCameraX = (2 * ((x + 0.5) / _resolution.x()) - 1) * aspectRatio * scale;
