+++ src/Camera.h
@@ -13,14 +13,14 @@ class Camera
 private:
   Vector3d _origin;
   float    _fov; // TODO: relate to alpha
-  Vector2d _resolution; // (width, height)
+  Vector2i _resolution; // (width, height)
 
 public:
-  Camera( const Vector3d &origin, const float &fov, const Vector2d &resolution );
+  Camera( const Vector3d &origin, const float &fov, const Vector2i &resolution );
 
   std::map< std::vector<int>, Ray > getAllRays(); // Function for precomputing all rays
 
-  Vector2d getResolution()
+  Vector2i getResolution()
   {
   	return _resolution;
   }
